+++ packages/orq-rc/RELEASES.md
@@ -1409,3 +1409,13 @@ Based on:
 - [typescript v3.4.0-rc.28] packages/orq-rc
 ### Releases
 - [NPM v3.4.0-rc.28] https://www.npmjs.com/package/@orq-ai/node/v/3.4.0-rc.28 - packages/orq-rc
+
+## 2025-04-16 11:27:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.533.0 (2.578.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.4.0-rc.29] packages/orq-rc
+### Releases
+- [NPM v3.4.0-rc.29] https://www.npmjs.com/package/@orq-ai/node/v/3.4.0-rc.29 - packages/orq-rc